Fixed Amplitude Modulation in FM voices
Looking at multiple sources, including an official Yamaha YM2608 manual,
the only bit relevant to AM is the high bit. Bits 5 and 6 are completely
unused.

Because this is yet another change to the SMPS2ASM format, there is a
check that retains backward compatibility with older files.

Notably, a voice in S2's credits has this bit set, though it has no
effect because S2's driver doesn't support LFO.

I'll have to modify the SMPS2ASM converter tool to follow this
behaviour.