Support S3 and S&K DAC layouts.
While S&K worked with the Flamedriver logic, support for Sonic 3
was broken since the Flamedriver logic results in an alternate DAC
order.

This does assume that setting 'SonicDriverVer' to 3 means Sonic 3 and
that 4 means Sonic & Knuckles. If Sonic 3D Blast is ever supported
like this, then it can be given 6.